% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipality.R
\name{kld_municipalities}
\alias{kld_municipalities}
\title{List municipalities from the Kolada API}
\usage{
kld_municipalities()
}
\value{
A \code{tibble} with columns:
\describe{
\item{id}{The unique identifier for the municipality or county (e.g., "0180" for Linköping).}
\item{title}{The name of the municipality or county (e.g., "Linköping").}
\item{type}{The type of the geographical entity ("county" or "municipality").}
}
}
\description{
This function retrieves a list of all municipalities and counties available
in the Kolada database.
}
\examples{
# Get a list of all municipalities and counties
all_municipalities <- kld_municipalities()
head(all_municipalities)
tail(all_municipalities)
}
